requirements.txt
pandas
pydantic
firecrawl-py
google-genai
python-dotenv
tabulate
